+++ Neuron.m
@@ -58,7 +58,7 @@ classdef Neuron < sbfsem.core.NeuronAPI
             %   % Import c127 in NeitzInferiorMonkey
             %   c127 = Neuron(127, 'i');
             % -------------------------------------------------------------
-            obj@sbfsem.core.NeuronAPI();
+            obj@sbfsem.core.NeuronAPI(ID, source);
 
             % By default, synapses are not imported
             if nargin < 3
@@ -75,37 +75,11 @@ classdef Neuron < sbfsem.core.NeuronAPI
                 obj.transform = sbfsem.core.Transforms.fromStr(transform);
             end
 
-            % NeuronCache inputs a single cell, cmd line as 2-3 arguments
-            if nargin == 1 && iscell(ID)
-                source = ID{2};
-                if numel(ID) > 2
-                    validateattributes(ID{3}, {'logical'}, {});
-                    obj.includeSynapses = ID{3};
-                end
-                ID = ID{1};
-            elseif nargin == 3
-                validateattributes(includeSynapses, {'logical'}, {});
-                obj.includeSynapses = includeSynapses;
-            elseif nargin == 4
-                if isa(transform, 'sbfsem.core.Transforms')
-                    obj.transform = transform;
-                else
-                    obj.transform = sbfsem.core.Transforms.fromStr(transform);
-                end
-            end
-
-            validateattributes(ID, {'numeric'}, {'numel', 1});
             source = validateSource(source);
             obj.transform = validateTransform(obj.transform, source);
-            obj.ID = ID;
-            obj.source = source;
             fprintf('-----c%u-----\n', obj.ID);
 
-            % XYZ volume dimensions in nm/pix, nm/pix, nm/sections
-            obj.volumeScale = getODataScale(obj.source);
-
             % Instantiate OData clients
-            obj.ODataClient = sbfsem.io.NeuronOData(obj.ID, obj.source);
             if obj.includeSynapses
                 obj.SynapseClient = sbfsem.io.SynapseOData(obj.ID, obj.source);
             else
@@ -116,8 +90,6 @@ classdef Neuron < sbfsem.core.NeuronAPI
             % Fetch neuron OData and parse
             obj.pull();
 
-            % Track when the Neuron object was created
-            obj.lastModified = datestr(now);
             fprintf('\n\n');
         end
 
@@ -171,67 +143,19 @@ classdef Neuron < sbfsem.core.NeuronAPI
             end
         end
     end
+    
+    methods (Access = protected)
+                
+        function pull(obj)
+            pull@sbfsem.core.StructureAPI(obj)
+            if obj.includeSynapses
+                obj.getSynapses();
+            end
+        end
+    end
 
     methods (Access = private)
-        function pull(obj)
-            % PULL  Fetch and parse neuron's OData
 
-            % Get the relevant data with OData queries
-            [obj.viking, obj.nodes, obj.edges] = obj.ODataClient.pull();
-            % XY transform and then convert data to microns
-            obj.nodes = obj.setXYZum(obj.nodes);
-            
-            if obj.includeSynapses
-                [obj.synapses, childNodes, childEdges] = obj.SynapseClient.pull();
-                obj.nodes = [obj.nodes; obj.setXYZum(childNodes)];
-                obj.edges = [obj.edges; childEdges];
-                obj.setupSynapses();
-            end
-            
-            if nnz(obj.nodes.Geometry == 6)
-                obj.getGeometries();
-                fprintf('     %u closed curve geometries\n',...
-                    height(obj.geometries));
-            end
-                        
-            % Search for omitted nodes by location ID and section number
-            obj.omittedIDs = omitLocations(obj.ID, obj.source);
-            omittedSections = omitSections(obj.source);
-            if ~isempty(omittedSections)
-                for i = 1:numel(omittedSections)
-                    row = obj.nodes.Z == omittedSections(i);
-                    obj.omittedIDs = [obj.omittedIDs; obj.nodes(row,:).ID];
-                end
-            end
-        end
-
-        function nodes = setXYZum(obj, nodes)
-            % SETXYZUM  Convert Viking pixels to microns
-            if nnz(nodes.X) + nnz(nodes.Y) > 2
-                nodes = estimateSynapseXY(obj, nodes);
-            end
-            
-            % Apply transforms to NeitzInferiorMonkey
-            if obj.transform == sbfsem.core.Transforms.SBFSEMTools
-                xyDir = [fileparts(mfilename('fullpath')), '\data'];
-                xydata = dlmread([xyDir,...
-                    '\XY_OFFSET_NEITZINFERIORMONKEY.txt']);
-                volX = nodes.X + xydata(nodes.Z,2);
-                volY = nodes.Y + xydata(nodes.Z,3);
-            else
-                volX = nodes.VolumeX;
-                volY = nodes.VolumeY;
-            end
-
-            % Create an XYZ in microns column
-            nodes.XYZum = zeros(height(nodes), 3);
-            % TODO: There's an assumption about the units in here...
-            nodes.XYZum = bsxfun(@times,...
-                [volX, volY, nodes.Z],...
-                (obj.volumeScale./1e3));
-            % Create a column for radius in microns
-            nodes.Rum = nodes.Radius * obj.volumeScale(1)./1000;
-        end
 
         function setupSynapses(obj)
             % SETUPSYNAPSES